+++ multi-step-data-collection/src/components/register/BaseFormPage.tsx
@@ -1,3 +1,4 @@
+// BaseFormPage.tsx
 import React, { ReactNode } from 'react';
 import LinearProgress from '@mui/material/LinearProgress';
